// src/V2/types.ts
export interface emoteResponse {
    response: string;
    status: "success" | "fancySuccess" | "error" | "info" | "neutral" | "error";
}

export type IHyperion = V2Type;
// ...
export interface ICommandContext<T = Module>{
    msg: Message;
    channel: TextChannel;
    guild: Guild;
    guildConfig: GuildType;
    member: Member;
    user: User;
    command: Command;
    module: T;
    content: string;
    args: Array<string>;
    dev: boolean;
    admin: boolean;
    override?: boolean;
    permLevel?: number;
    respond?: string;
    silent?: boolean;
    delete?: boolean;
    mod: boolean;
    createMessage(input: MessageContent): void;
}